move App.js to react/App.js
